+++ src/API.mli
@@ -245,7 +245,8 @@ module Pp : sig
   val constraints : Data.pretty_printer_context -> Format.formatter -> Data.constraints -> unit
   val state : Format.formatter -> Data.state -> unit
 
-  val query : Format.formatter -> 'a Compile.query -> unit
+  val program : Format.formatter -> 'a Compile.query -> unit
+  val goal : Format.formatter -> 'a Compile.query -> unit
 
   module Ast : sig
     val program : Format.formatter -> Ast.program -> unit